python: Rework wrapper
This adds a layer of python on top of the C library wrapper.
Historically the wrapper was kept to a pure C implementation to allow
autotools/cmake to build the library, cli and python wrapper.
There are however too many limitations with that approach and a more
standard python setuptools configuration was added.

With that in place it makes more sense to move some stuff from the C
wrapper.

With this commit the C wrapper mainly deals with byte arrays and most of
the string handling, UUID conversion, checksum generation are moved to
the python top layer.

--- New functionallity ---
Session.authenticate_dsa_token - New, exposes the underlying DSA token authentication mechanism
Session.part_erase             - Erase a partition

--- Breaking changes ---
part_get_partitions - New, Replaces 'part_list_partitions' which only
                    exposed UUID's and description fiellds of partitions.
                    This new function returns all information about partitions.

--- Updated functions that does not break API compability ---
__init__     - Implemented optional UUID selection parameter, with this
it's now possible to initialize sessions to different puchboot devices
attached to the same host.

part_verify  - Can now accept either a pathlib Path, a byte array or a BufferedReader,
                and part argument can be either a string version of an uuid or an UUID object
part_write   - Can now accept either a pathlib Path or a BufferedReader
                and part argument can be either a string version of an uuid or an UUID object
part_table_install - Can now accept string or UUID object for part parameter

board_run_command - Can now accept both command ID's and string version of the command,
                    Support for passing an optinal byte array as argument has also been added
                    The underlying API has changed slightly to support returning byte arrays,
                    and the python does the same thing now.

--- Deprecations ---
part_resize - This has not been supported for a while in the bootloader, and
                now it has been removed from punchboot tools as well