cli: part: Remove 'offset' parameter for write/verify
The offset parameter on verify was never used. On write the offset was
ignored if it's a bpak file. In addition the option was not properply
implemented in getopts and not present in the help output.

Removed for now. Maybe there will be support for writing/verifying stuff
at an offset in the future.